requirements.txt
google-auth
google-generativeai
gspread
jinja2
python-dotenv
streamlit
weasyprint
pandas
openai
oauth2client
pytest
flake8
black
pytest-mock